+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/delete_key_operation.hpp
@@ -78,6 +78,12 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
 
   public:
     /**
+     * @brief Get the #Azure::Core::Http::RawResponse of the operation request.
+     * @return A pointer to #Azure::Core::Http::RawResponse or null.
+     */
+    Azure::Core::Http::RawResponse* GetRawResponse() const override { return m_rawResponse.get(); }
+
+    /**
      * @brief Get the #Azure::Security::KeyVault::Keys::DeletedKey object.
      *
      * @remark The deleted key contains the recovery id if the key can be recovered.